working text to command code read

+++ include/randomstuff.h
@@ -34,9 +34,9 @@
 
 #define gAngle -robit->p.mRot * (PI / 180)
 #define renderRad 1//amount of the robot's radii that are used to calculate cone distance, smaller is more optimized (but calculates for less cones)
-inline int getSign(float value) {//returns whether a number is negative or positive.
-	if (value < 0) { return -1; }
-	else if (value > 0) { return 1; }
+inline float getSign(float value) {//returns whether a number is negative or positive.
+	if (value < 0.0) { return -1; }
+	else if (value > 0.0) { return 1; }
 	else { return 0; }
 }
 inline float sqr(float value) {
